+++ recipes/accessing_a_list_of_elements.py
@@ -35,7 +35,7 @@ def main():
     # See how complex is our wikipedia documents:
     print("number of headings:", len(body.headers))
     print("number of images stored:", len(body.images))
-    print("number of paragraphs:", len(body.get_paragraphs()))
+    print("number of paragraphs:", len(body.paragraphs))
     print("number of links (URLs):", len(body.get_links()))
     print("number of footnotes:", len(body.get_notes()))
     # Our sample document has no table:
@@ -76,7 +76,7 @@ def test_unit(document: Document) -> None:
     assert count_methods == 96
     assert len(body.headers) == 29
     assert len(body.images) == 0
-    assert len(body.get_paragraphs()) == 175
+    assert len(body.paragraphs) == 175
     assert len(body.get_links()) == 352
     assert len(body.get_notes()) == 49
     assert len(body.get_paragraphs(content=r"Fish")) == 4